time variables are double precision

+++ Program.cs
@@ -7,9 +7,9 @@ class Program
 {
     static void Main()
     {
-        float t = 0.0f;    // time
-        float tEnd = 3.0f; // ending time
-        float dt = 0.02f;  // time step
+        double t = 0.0;    // time
+        double tEnd = 3.0; // ending time
+        double dt = 0.02;  // time step
 
         // time loop
         while(t < tEnd)
